Added save/editing for Parameter

+++ src/main/java/com/parameters/prts/Repository/ParameterRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ParameterRepository extends JpaRepository<ParameterEntity, Long> {
+public interface ParameterRepository extends JpaRepository<ParameterEntity, Integer> {
     List<ParameterEntity> findByVidContaining(String title);
 }
